@@ -4,5 +4,6 @@ public record UpdateProfileRequest(
         String name,
         String surname,
         String avatarUrl,
-        String level) {
+        String level
+) {
 }
