@@ -1,6 +1,8 @@
 package com.mockly.user.service;
 
 import com.mockly.user.dto.AuditAction;
+import com.mockly.user.dto.AuditEventMessage;
+import com.mockly.user.dto.AuditProperties;
 import lombok.Builder;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
@@ -16,39 +18,30 @@ import java.util.UUID;
 public class AuditEventPublisher {
 
     private final RabbitTemplate rabbitTemplate;
-    private static final String AUDIT_EXCHANGE = "audit.exchange";
-    private static final String AUDIT_ROUTING_KEY = "audit.event";
-
-    @Data
-    @Builder
-    public static class AuditEvent {
-        private UUID userId;
-        private String role;
-        private AuditAction action;
-        private String entityType;
-        private UUID entityId;
-        private String result;
-        private String details;
-    }
+    private final AuditProperties auditProperties;
 
     public void publishEvent(UUID userId, String role, AuditAction action,
-            String entityType, UUID entityId, String result, String details) {
+                             String entityType, UUID entityId, String result, String details) {
         try {
-            AuditEvent event = AuditEvent.builder()
-                    .userId(userId)
-                    .role(role)
-                    .action(action)
-                    .entityType(entityType)
-                    .entityId(entityId)
-                    .result(result)
-                    .details(details)
-                    .build();
+            AuditEventMessage event = new AuditEventMessage(
+                    userId,
+                    role,
+                    action,
+                    entityType,
+                    entityId,
+                    result,
+                    details
+            );
 
-            // Fire and forget event to RabbitMQ
-            rabbitTemplate.convertAndSend(AUDIT_EXCHANGE, AUDIT_ROUTING_KEY, event);
-            log.debug("Published audit event: {} for {}", action, userId);
+            rabbitTemplate.convertAndSend(
+                    auditProperties.getExchange(),
+                    auditProperties.getRoutingKey(),
+                    event
+            );
+
+            log.debug("Published audit event: action={}, userId={}", action, userId);
         } catch (Exception e) {
-            log.error("Failed to publish audit event for {}", userId, e);
+            log.error("Failed to publish audit event for userId={}", userId, e);
         }
     }
 }
